Refactor database code into models
